+++ kirlent_docutils/__init__.py
@@ -5,7 +5,7 @@
 
 """Custom writers for docutils."""
 
-__version__ = "0.2"
+__version__ = "0.3"
 
 from docutils.core import publish_cmdline